# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Revoked API
  version: "0.1.0"
  summary: Secret sharing and data collection over revocable, living grants.
  description: |
    Every operator runs this API under their own domain, so there is no shared
    base URL — substitute your own host everywhere below.

    Two things distinguish this API from a typical CRUD surface, and both are
    worth reading before integrating:

    - **Grants are living.** A share resolves the current value at read time,
      so rotating a secret updates every share that points at it. Pausing or
      revoking one takes effect on the next read, not on the next sync.
    - **Public endpoints are protected by slug entropy alone.** Anything under
      `/api/public` or `/s` is reachable by whoever holds the slug. There is
      deliberately no list endpoint for them.

    Endpoints marked *unauthenticated* are meant to be called by people with no
    account on this server — recipients of a share, responders to a request,
    verifiers walking the trust chain.
  license:
    name: See repository
servers:
  - url: https://{host}
    description: Your deployment
    variables:
      host:
        default: api.example.com
        description: The domain this operator runs under, matching its DNS TXT record.
security:
  - ApiKey: []
  - Session: []
...
paths:
...
  /passkey:
    get:
      tags: [Sign-in]
      summary: The sign-in page
      description: |
        The page every passkey ceremony runs on, opened in the system browser.
        What it does is read from its address: `mode=signin` (the default) or
        `mode=signup`, or a `ticket` to add a passkey to the account the
        ticket names. An app passes its PKCE `challenge` and a `state`; when
        the ceremony succeeds the page opens
        `revoked://auth?code=…&state=…`, and the app redeems the code at
        `/api/passkeys/token`.
      security: []
      parameters:
        - { name: mode, in: query, schema: { type: string, enum: [signin, signup, enroll] } }
        - { name: ticket, in: query, schema: { type: string } }
        - { name: challenge, in: query, schema: { type: string }, description: "PKCE S256 challenge, base64url." }
        - { name: state, in: query, schema: { type: string } }
      responses:
        "200":
          description: The page.
          content:
            text/html: {}
  /api/passkeys/login/begin:
    post:
      tags: [Sign-in]
      summary: Start a passkey sign-in
      description: |
        Returns WebAuthn request options for a discoverable credential — no
        account is named; the authenticator's answer says whose passkey it is.
        Must be called at the server's own address (`localhost` on a
        development machine, never a loopback IP). Rate-limited per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge]
              properties:
                challenge: { type: string, description: "The app's PKCE S256 challenge; the code issued at the end is bound to it." }
      responses:
        "200":
          description: The ceremony, good for five minutes.
          content:
            application/json:
              schema:
                type: object
                properties:
                  session: { type: string }
                  options: { type: object, description: "`{publicKey: PublicKeyCredentialRequestOptions}`, binary members base64url." }
        "400":
          description: "`passkey_request_invalid` or `passkey_origin_invalid`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/passkeys/login/finish:
    post:
      tags: [Sign-in]
      summary: Finish a passkey sign-in
      description: |
        Verifies the authenticator's assertion and returns a one-time code,
        good for two minutes, that only the holder of the PKCE verifier can
        redeem. A ceremony is answered once, right or wrong.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session, credential]
              properties:
                session: { type: string }
                credential: { type: object, description: "The `PublicKeyCredential`, binary members base64url." }
      responses:
        "200":
          description: The one-time code.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string }
        "400":
          description: "`passkey_ceremony_invalid`, `passkey_request_invalid` or `passkey_origin_invalid`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
        "401":
          description: "`passkey_verification_failed`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/passkeys/register/begin:
    post:
      tags: [Sign-in]
      summary: Start registering a passkey
      description: |
        Either a new account — `email`, where the operator allows signups — or
        a `ticket`, which adds a passkey to the account it names. Returns
        WebAuthn creation options that require a discoverable credential and
        user verification. Nothing is written until the ceremony finishes.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, description: "The new account's address. Ignored with a ticket." }
                ticket: { type: string, description: "A one-time ticket from `/api/passkeys/tickets` or the operator." }
                name: { type: string, maxLength: 60, description: "What the passkey is called in the owner's list." }
                challenge: { type: string, description: "The app's PKCE S256 challenge, when the result should sign an app in." }
      responses:
        "200":
          description: The ceremony, good for five minutes.
          content:
            application/json:
              schema:
                type: object
                properties:
                  session: { type: string }
                  options: { type: object, description: "`{publicKey: PublicKeyCredentialCreationOptions}`, binary members base64url." }
        "400":
          description: "`passkey_email_invalid`, `passkey_ticket_invalid`, `passkey_request_invalid` or `passkey_origin_invalid`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
        "403":
          description: "`signups_disabled`: no ticket, and this server does not accept registrations."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
        "409":
          description: "`passkey_email_taken`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/passkeys/register/finish:
    post:
      tags: [Sign-in]
      summary: Finish registering a passkey
      description: |
        Verifies the attestation, then writes the passkey — and, for a signup,
        the account — and spends the ticket. Returns a one-time code when the
        ceremony was started with a PKCE challenge.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session, credential]
              properties:
                session: { type: string }
                credential: { type: object, description: "The `PublicKeyCredential`, binary members base64url." }
      responses:
        "200":
          description: Saved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string, description: "Present when a PKCE challenge was given." }
        "400":
          description: "`passkey_ceremony_invalid`, `passkey_verification_failed`, `passkey_ticket_invalid` or `passkey_origin_invalid`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
        "409":
          description: "`passkey_email_taken`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/passkeys/token:
    post:
      tags: [Sign-in]
      summary: Exchange a code for a session (the app)
      description: |
        Spends the one-time code — right or wrong, a code works once — and,
        when the verifier matches its challenge, returns a session token. Not
        needed for API-key access.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, verifier]
              properties:
                code: { type: string }
                verifier: { type: string, description: "The PKCE verifier." }
      responses:
        "200":
          description: A session token and the user record.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  record: { type: object }
        "400":
          description: "`passkey_grant_invalid`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/passkeys/tickets:
    post:
      tags: [Sign-in]
      summary: Issue a link to add a passkey
      description: |
        Returns a one-time link, good for fifteen minutes, to the sign-in page
        where the caller's account can register another passkey — on this
        device or, opened elsewhere, on another. Users only: an API key cannot
        let itself in as the person behind it.
      security:
        - Session: []
      responses:
        "200":
          description: The link.
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }
                  expiresAt: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/AppError" }
  /api/passkeys/{id}:
    delete:
      tags: [Sign-in]
      summary: Remove a passkey
      description: |
        Removes one of the caller's passkeys; it stops signing in at once. The
        last one cannot be removed. Passkeys are listed through
        `/api/collections/passkeys/records` (name, created, last used — never
        the key).
      security:
        - Session: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Removed. }
        "401": { $ref: "#/components/responses/AppError" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`passkey_last`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Programmatic access. Sending a key as a Bearer token authenticates as
        nobody — the request is treated as a guest and refused by the rules.
    Session:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A session token from sign-in. Used by the app.
    Connection:
      type: apiKey
      in: header
      name: X-Revoked-Connection
      description: A connected tool's token, from `/api/connect/token`.
...
  responses:
    NotFound:
      description: No such resource, or the caller may not see it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AppError" }
    AppError:
      description: Refused, with a named reason.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AppError" }
    RateLimited:
      description: |
        Too many attempts. Password gates, probes and challenges are limited
        per client, so a slug cannot be brute-forced.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AppError" }
  schemas:
    AppError:
      type: object
      description: |
        The stable error envelope. Key off `code`, never off `message` — the
        prose may change, the code will not.

        Messages may echo what the caller sent, never server-held state: an
        error outlives the credential that produced it and lands in logs and
        screenshots read by people who hold no key.
      properties:
        code: { type: string, examples: ["link_password_required"] }
        message: { type: string }
        status: { type: integer }
...